Ensure iterated and emitted types are registered.
