+++ sdks/go/pkg/beam/register/register_test.go
@@ -21,13 +21,10 @@ import (
 	"testing"
 
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/graph"
-	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/graph/mtime"
-	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/graph/window"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/runtime"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/runtime/exec"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/runtime/graphx"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/runtime/graphx/schema"
-	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/typex"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/util/reflectx"
 )
 
@@ -321,219 +318,6 @@ func TestCombiner_PartialCombiner2(t *testing.T) {
 	}
 	if got, want := ma.Call([]interface{}{2, 4})[0].(int), 6; got != want {
 		t.Errorf("Wrapped MergeAccumulators call: got %v, want %v", got, want)
-	}
-}
-
-func TestEmitter1(t *testing.T) {
-	Emitter1[int]()
-	if !exec.IsEmitterRegistered(reflect.TypeOf((*func(int))(nil)).Elem()) {
-		t.Fatalf("exec.IsEmitterRegistered(reflect.TypeOf((*func(int))(nil)).Elem()) = false, want true")
-	}
-}
-
-func TestEmitter2(t *testing.T) {
-	Emitter2[int, string]()
-	if !exec.IsEmitterRegistered(reflect.TypeOf((*func(int, string))(nil)).Elem()) {
-		t.Fatalf("exec.IsEmitterRegistered(reflect.TypeOf((*func(int, string))(nil)).Elem()) = false, want true")
-	}
-}
-
-func TestEmitter2_WithTimestamp(t *testing.T) {
-	Emitter2[typex.EventTime, string]()
-	if !exec.IsEmitterRegistered(reflect.TypeOf((*func(typex.EventTime, string))(nil)).Elem()) {
-		t.Fatalf("exec.IsEmitterRegistered(reflect.TypeOf((*func(typex.EventTime, string))(nil)).Elem()) = false, want true")
-	}
-}
-
-func TestEmitter3(t *testing.T) {
-	Emitter3[typex.EventTime, int, string]()
-	if !exec.IsEmitterRegistered(reflect.TypeOf((*func(typex.EventTime, int, string))(nil)).Elem()) {
-		t.Fatalf("exec.IsEmitterRegistered(reflect.TypeOf((*func(typex.EventTime, int, string))(nil)).Elem()) = false, want true")
-	}
-}
-
-func TestEmit1(t *testing.T) {
-	e := &emit1[int]{n: &elementProcessor{}}
-	e.Init(context.Background(), []typex.Window{}, mtime.ZeroTimestamp)
-	fn := e.Value().(func(int))
-	fn(3)
-	if got, want := e.n.(*elementProcessor).inFV.Elm, 3; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Elm=%v, want %v", got, want)
-	}
-	if got := e.n.(*elementProcessor).inFV.Elm2; got != nil {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Elm2=%v, want nil", got)
-	}
-	if got, want := e.n.(*elementProcessor).inFV.Timestamp, mtime.ZeroTimestamp; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Timestamp=%v, want %v", got, want)
-	}
-}
-
-func TestEmit2(t *testing.T) {
-	e := &emit2[int, string]{n: &elementProcessor{}}
-	e.Init(context.Background(), []typex.Window{}, mtime.ZeroTimestamp)
-	fn := e.Value().(func(int, string))
-	fn(3, "hello")
-	if got, want := e.n.(*elementProcessor).inFV.Elm, 3; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Elm=%v, want %v", got, want)
-	}
-	if got, want := e.n.(*elementProcessor).inFV.Elm2, "hello"; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Elm2=%v, want %v", got, want)
-	}
-	if got, want := e.n.(*elementProcessor).inFV.Timestamp, mtime.ZeroTimestamp; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Timestamp=%v, want %v", got, want)
-	}
-}
-
-func TestEmit1WithTimestamp(t *testing.T) {
-	e := &emit1WithTimestamp[int]{n: &elementProcessor{}}
-	e.Init(context.Background(), []typex.Window{}, mtime.ZeroTimestamp)
-	fn := e.Value().(func(typex.EventTime, int))
-	fn(mtime.MaxTimestamp, 3)
-	if got, want := e.n.(*elementProcessor).inFV.Elm, 3; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Elm=%v, want %v", got, want)
-	}
-	if got := e.n.(*elementProcessor).inFV.Elm2; got != nil {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Elm2=%v, want nil", got)
-	}
-	if got, want := e.n.(*elementProcessor).inFV.Timestamp, mtime.MaxTimestamp; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Timestamp=%v, want %v", got, want)
-	}
-}
-
-func TestEmit2WithTimestamp(t *testing.T) {
-	e := &emit2WithTimestamp[int, string]{n: &elementProcessor{}}
-	e.Init(context.Background(), []typex.Window{}, mtime.ZeroTimestamp)
-	fn := e.Value().(func(typex.EventTime, int, string))
-	fn(mtime.MaxTimestamp, 3, "hello")
-	if got, want := e.n.(*elementProcessor).inFV.Elm, 3; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Elm=%v, want %v", got, want)
-	}
-	if got, want := e.n.(*elementProcessor).inFV.Elm2, "hello"; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Elm2=%v, want %v", got, want)
-	}
-	if got, want := e.n.(*elementProcessor).inFV.Timestamp, mtime.MaxTimestamp; got != want {
-		t.Errorf("e.Value.(func(int))(3).n.inFV.Timestamp=%v, want %v", got, want)
-	}
-}
-
-func TestIter1(t *testing.T) {
-	Iter1[int]()
-	if !exec.IsInputRegistered(reflect.TypeOf((*func(*int) bool)(nil)).Elem()) {
-		t.Fatalf("exec.IsInputRegistered(reflect.TypeOf(((*func(*int) bool)(nil)).Elem()) = false, want true")
-	}
-}
-
-func TestIter2(t *testing.T) {
-	Iter2[int, string]()
-	if !exec.IsInputRegistered(reflect.TypeOf((*func(*int, *string) bool)(nil)).Elem()) {
-		t.Fatalf("exec.IsInputRegistered(reflect.TypeOf((*func(*int, *string) bool)(nil)).Elem()) = false, want true")
-	}
-}
-
-func TestIter1_Struct(t *testing.T) {
-	values := []exec.FullValue{exec.FullValue{
-		Windows:   window.SingleGlobalWindow,
-		Timestamp: mtime.ZeroTimestamp,
-		Elm:       "one",
-	}, exec.FullValue{
-		Windows:   window.SingleGlobalWindow,
-		Timestamp: mtime.ZeroTimestamp,
-		Elm:       "two",
-	}, exec.FullValue{
-		Windows:   window.SingleGlobalWindow,
-		Timestamp: mtime.ZeroTimestamp,
-		Elm:       "three",
-	}}
-
-	i := iter1[string]{s: &exec.FixedReStream{Buf: values}}
-
-	i.Init()
-	fn := i.Value().(func(value *string) bool)
-
-	var s string
-	if ok := fn(&s); !ok {
-		t.Fatalf("First i.Value()(&s)=false, want true")
-	}
-	if got, want := s, "one"; got != want {
-		t.Fatalf("First iter value = %v, want %v", got, want)
-	}
-	if ok := fn(&s); !ok {
-		t.Fatalf("Second i.Value()(&s)=false, want true")
-	}
-	if got, want := s, "two"; got != want {
-		t.Fatalf("Second iter value = %v, want %v", got, want)
-	}
-	if ok := fn(&s); !ok {
-		t.Fatalf("Third i.Value()(&s)=false, want true")
-	}
-	if got, want := s, "three"; got != want {
-		t.Fatalf("Third iter value = %v, want %v", got, want)
-	}
-	if ok := fn(&s); ok {
-		t.Fatalf("Fourth i.Value()(&s)=true, want false")
-	}
-	if err := i.Reset(); err != nil {
-		t.Fatalf("i.Reset()=%v, want nil", err)
-	}
-}
-
-func TestIter2_Struct(t *testing.T) {
-	values := []exec.FullValue{exec.FullValue{
-		Windows:   window.SingleGlobalWindow,
-		Timestamp: mtime.ZeroTimestamp,
-		Elm:       1,
-		Elm2:      "one",
-	}, exec.FullValue{
-		Windows:   window.SingleGlobalWindow,
-		Timestamp: mtime.ZeroTimestamp,
-		Elm:       2,
-		Elm2:      "two",
-	}, exec.FullValue{
-		Windows:   window.SingleGlobalWindow,
-		Timestamp: mtime.ZeroTimestamp,
-		Elm:       3,
-		Elm2:      "three",
-	}}
-
-	i := iter2[int, string]{s: &exec.FixedReStream{Buf: values}}
-
-	i.Init()
-	fn := i.Value().(func(key *int, value *string) bool)
-
-	var s string
-	var key int
-	if ok := fn(&key, &s); !ok {
-		t.Fatalf("First i.Value()(&s)=false, want true")
-	}
-	if got, want := key, 1; got != want {
-		t.Fatalf("First iter key = %v, want %v", got, want)
-	}
-	if got, want := s, "one"; got != want {
-		t.Fatalf("First iter value = %v, want %v", got, want)
-	}
-	if ok := fn(&key, &s); !ok {
-		t.Fatalf("Second i.Value()(&s)=false, want true")
-	}
-	if got, want := key, 2; got != want {
-		t.Fatalf("Second iter key = %v, want %v", got, want)
-	}
-	if got, want := s, "two"; got != want {
-		t.Fatalf("Second iter value = %v, want %v", got, want)
-	}
-	if ok := fn(&key, &s); !ok {
-		t.Fatalf("Third i.Value()(&s)=false, want true")
-	}
-	if got, want := key, 3; got != want {
-		t.Fatalf("Third iter key = %v, want %v", got, want)
-	}
-	if got, want := s, "three"; got != want {
-		t.Fatalf("Third iter value = %v, want %v", got, want)
-	}
-	if ok := fn(&key, &s); ok {
-		t.Fatalf("Fourth i.Value()(&s)=true, want false")
-	}
-	if err := i.Reset(); err != nil {
-		t.Fatalf("i.Reset()=%v, want nil", err)
 	}
 }
